requirement.txt
pandas
requests
pydub
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
scikit-learn
torch
transformers
datasets
soundfile
numpy